+++ dune/localfunctions/lagrange/pk2d.hh
@@ -13,19 +13,8 @@
 #include <dune/localfunctions/common/localtoglobaladaptors.hh>
 #include <dune/localfunctions/lagrange/lagrangesimplex.hh>
 
-#warning This header is deprecated
-
 namespace Dune
 {
-
-  /** \todo Please doc me !
-
-      \deprecated This class is obsolete. Please use LagrangeSimplexLocalFiniteElement instead!
-   */
-  template<class D, class R, unsigned int k>
-  using Pk2DLocalFiniteElement
-    [[deprecated("use LagrangeSimplexLocalFiniteElement instead")]]
-    = LagrangeSimplexLocalFiniteElement<D,R,2,k>;
 
 
   //! Langrange finite element of arbitrary order on triangles