+++ dune/localfunctions/rannacherturek/rannachertureklocalinterpolation.hh
@@ -13,7 +13,6 @@
 #include <dune/geometry/referenceelements.hh>
 
 #include <dune/localfunctions/common/localbasis.hh>
-#include <dune/localfunctions/common/localinterpolation.hh>
 
 namespace Dune
 {
@@ -36,11 +35,9 @@ namespace Dune
 
   public:
     template< class F, class C >
-    void interpolate ( const F &ff, std::vector< C > &out ) const
+    void interpolate ( const F &f, std::vector< C > &out ) const
     {
       typedef typename Traits::DomainType DomainType;
-
-      auto&& f = Impl::makeFunctionWithCallOperator<DomainType>(ff);
 
       // get cubic reference element
       auto referenceElement = ReferenceElements< D, d >::cube();