+++ dune/localfunctions/raviartthomas/raviartthomas02d/raviartthomas02dlocalinterpolation.hh
@@ -9,7 +9,6 @@
 #include <array>
 #include <bitset>
 #include <vector>
-#include <dune/localfunctions/common/localinterpolation.hh>
 
 namespace Dune
 {
@@ -40,10 +39,9 @@ namespace Dune
     }
 
     template<typename F, typename C>
-    void interpolate (const F& ff, std::vector<C>& out) const
+    void interpolate (const F& f, std::vector<C>& out) const
     {
       // f gives v*outer normal at a point on the edge!
-      auto&& f = Impl::makeFunctionWithCallOperator<typename LB::Traits::DomainType>(ff);
 
       out.resize(3);