+++ dune/localfunctions/refined/refinedp0/refinedp0localinterpolation.hh
@@ -6,7 +6,6 @@
 #define DUNE_REFINED_P0_LOCALINTERPOLATION_HH
 
 #include <dune/localfunctions/refined/refinedp0/refinedp0localbasis.hh>
-#include <dune/localfunctions/common/localinterpolation.hh>
 
 namespace Dune
 {
@@ -38,10 +37,8 @@ namespace Dune
 
 
     template<typename F, typename C>
-    void interpolate (const F& ff, std::vector<C>& out) const
+    void interpolate (const F& f, std::vector<C>& out) const
     {
-      auto&& f = Impl::makeFunctionWithCallOperator<DT>(ff);
-
       out.resize(interpolationPoints_.size());
       for (size_t i = 0; i < out.size(); ++i)
       {
@@ -82,10 +79,8 @@ namespace Dune
 
 
     template<typename F, typename C>
-    void interpolate (const F& ff, std::vector<C>& out) const
+    void interpolate (const F& f, std::vector<C>& out) const
     {
-      auto&& f = Impl::makeFunctionWithCallOperator<DT>(ff);
-
       out.resize(interpolationPoints_.size());
       for (size_t i = 0; i < out.size(); ++i)
       {
@@ -146,10 +141,8 @@ namespace Dune
 
 
     template<typename F, typename C>
-    void interpolate (const F& ff, std::vector<C>& out) const
+    void interpolate (const F& f, std::vector<C>& out) const
     {
-      auto&& f = Impl::makeFunctionWithCallOperator<DT>(ff);
-
       out.resize(interpolationPoints_.size());
       for (size_t i = 0; i < out.size(); ++i)
       {